+++ puncturedfem/__init__.py
@@ -6,6 +6,7 @@ from . import antilap
 from . import d2n
 from . import mesh
 from . import nystrom
+from . import poly
 from . import plot
 from . import quad
 from . import test